back-end/pest-classification/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
tensorflow==2.19.0
python-multipart==0.0.20
pillow==11.2.1
numpy==2.1.0